osx leopard port, fixed npot texture handling, major macos extern cleanup, build system extended
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@16782 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/sdlengine/SDLWindow.cpp
@@ -878,7 +878,7 @@ SDLWindow::setSwapInterval(unsigned theInterval)
     }
 #endif
 #ifdef AC_USE_OSX_CGL
-    const long myInterval = theInterval;  
+    const GLint myInterval = theInterval;  
     CGLError myError = CGLSetParameter(CGLGetCurrentContext(), kCGLCPSwapInterval, &myInterval);
     if (myError != 0) {
         AC_WARNING << "Cannot set swap interval, error=" << CGLErrorString(myError) << ", " <<
